requirements.txt
OpenAI
flask
flask-cors
google-cloud-storage
pg8000
sqlalchemy
cloud-sql-python-connector